requirements.txt
aiogram
python-dotenv
aiohttp
beautifulsoup4
